requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
archivessnake==0.9.1
    # via -r requirements.in
attrs==22.1.0
    # via archivessnake
boltons==21.0.0
    # via archivessnake
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via
    #   cul-archives-utils
    #   requests
charset-normalizer==2.1.0
    # via
    #   cul-archives-utils
    #   requests
cul-archives-utils @ git+https://github.com/cul/cul-archives-utils.git
    # via -r requirements.in
freezegun==1.2.1
    # via -r requirements.in
google-api-core==2.8.2
    # via google-api-python-client
google-api-python-client==2.55.0
    # via -r requirements.in
google-auth==2.9.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.3
    # via
    #   cul-archives-utils
    #   requests
jarowinkler==1.2.0
    # via rapidfuzz
more-itertools==8.13.0
    # via archivessnake
oauth2client==4.1.3
    # via -r requirements.in
protobuf==4.21.4
    # via google-api-core
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   freezegun
pyyaml==6.0
    # via archivessnake
rapidfuzz==2.4.2
    # via archivessnake
requests==2.28.1
    # via
    #   archivessnake
    #   cul-archives-utils
    #   google-api-core
    #   responses
responses==0.21.0
    # via cul-archives-utils
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
structlog==22.1.0
    # via archivessnake
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.10
    # via
    #   cul-archives-utils
    #   requests
    #   responses
